Fixed application error
The web app wasn’t designed to easily handle changing the graduation
years and crashes if the range is outside 2016-2019. A temporary fix is
to keep the values in this range but this is not a good permanent
solution. A permanent solution is to redesign this to handle dynamic
date ranges.